+++ app/kanban-elastico/page.tsx
@@ -1,0 +1,5 @@
+import KanbanElasticoPrototype from './prototype';
+
+export default function KanbanElasticoPage() {
+  return <KanbanElasticoPrototype />;
+}